feat: add default values for users #682
- moved long constants to const for display names
- created functions to update new users with name, image, displayname, and username
- updated jwt callback
- updated create personal team and new team to include a team image too
- minor tweaks to team switcher component

Closes #682
