+++ apps/ai/src/lib/const.ts
@@ -6,3 +6,228 @@ export const SITE_URL =
   process.env.NEXT_PUBLIC_VERCEL_ENV === 'preview'
     ? `https://${process.env.NEXT_PUBLIC_VERCEL_URL}`
     : process.env.NEXT_PUBLIC_APP_URL ?? VERCEL_URL ?? 'http://localhost:3001';
+
+export const pokemonNames: string[] = [
+  'Bulbasaur',
+  'Ivysaur',
+  'Venusaur',
+  'Charmander',
+  'Charmeleon',
+  'Charizard',
+  'Squirtle',
+  'Wartortle',
+  'Blastoise',
+  'Caterpie',
+  'Metapod',
+  'Butterfree',
+  'Weedle',
+  'Kakuna',
+  'Beedrill',
+  'Pidgey',
+  'Pidgeotto',
+  'Pidgeot',
+  'Rattata',
+  'Raticate',
+  'Spearow',
+  'Fearow',
+  'Ekans',
+  'Arbok',
+  'Pikachu',
+  'Raichu',
+  'Sandshrew',
+  'Sandslash',
+  'Nidoran',
+  'Nidorina',
+  'Nidoqueen',
+  'Nidoran',
+  'Nidorino',
+  'Nidoking',
+  'Clefairy',
+  'Clefable',
+  'Vulpix',
+  'Ninetales',
+  'Jigglypuff',
+  'Wigglytuff',
+  'Zubat',
+  'Golbat',
+  'Oddish',
+  'Gloom',
+  'Vileplume',
+  'Paras',
+  'Parasect',
+  'Venonat',
+  'Venomoth',
+  'Diglett',
+  'Dugtrio',
+  'Meowth',
+  'Persian',
+  'Psyduck',
+  'Golduck',
+  'Mankey',
+  'Primeape',
+  'Growlithe',
+  'Arcanine',
+  'Poliwag',
+  'Poliwhirl',
+  'Poliwrath',
+  'Abra',
+  'Kadabra',
+  'Alakazam',
+  'Machop',
+  'Machoke',
+  'Machamp',
+  'Bellsprout',
+  'Weepinbell',
+  'Victreebel',
+  'Tentacool',
+  'Tentacruel',
+  'Geodude',
+  'Graveler',
+  'Golem',
+  'Ponyta',
+  'Rapidash',
+  'Slowpoke',
+  'Slowbro',
+  'Magnemite',
+  'Magneton',
+  'Farfetched',
+  'Doduo',
+  'Dodrio',
+  'Seel',
+  'Dewgong',
+  'Grimer',
+  'Muk',
+  'Shellder',
+  'Cloyster',
+  'Gastly',
+  'Haunter',
+  'Gengar',
+  'Onix',
+  'Drowzee',
+  'Hypno',
+  'Krabby',
+  'Kingler',
+  'Voltorb',
+  'Electrode',
+  'Exeggcute',
+  'Exeggutor',
+  'Cubone',
+  'Marowak',
+  'Hitmonlee',
+  'Hitmonchan',
+  'Lickitung',
+  'Koffing',
+  'Weezing',
+  'Rhyhorn',
+  'Rhydon',
+  'Chansey',
+  'Tangela',
+  'Kangaskhan',
+  'Horsea',
+  'Seadra',
+  'Goldeen',
+  'Seaking',
+  'Staryu',
+  'Starmie',
+  'Mr. Mime',
+  'Scyther',
+  'Jynx',
+  'Electabuzz',
+  'Magmar',
+  'Pinsir',
+  'Tauros',
+  'Magikarp',
+  'Gyarados',
+  'Lapras',
+  'Ditto',
+  'Eevee',
+  'Vaporeon',
+  'Jolteon',
+  'Flareon',
+  'Porygon',
+  'Omanyte',
+  'Omastar',
+  'Kabuto',
+  'Kabutops',
+  'Aerodactyl',
+  'Snorlax',
+  'Articuno',
+  'Zapdos',
+  'Moltres',
+  'Dratini',
+  'Dragonair',
+  'Dragonite',
+  'Mewtwo',
+  'Mew',
+];
+
+export const colorNames: string[] = [
+  'Red',
+  'Blue',
+  'Green',
+  'Yellow',
+  'Orange',
+  'Purple',
+  'Magenta',
+  'Turquoise',
+  'Teal',
+  'Lime',
+  'Coral',
+  'Aqua',
+  'Mauve',
+  'Lavender',
+  'Pink',
+  'Maroon',
+  'Olive',
+  'Navy',
+  'Grey',
+  'Tan',
+  'Crimson',
+  'Fuchsia',
+  'Cyan',
+  'Indigo',
+  'Violet',
+  'Cerise',
+  'Rust',
+  'Gold',
+  'Silver',
+  'Bronze',
+  'Amber',
+  'Peach',
+  'Mint',
+  'Emerald',
+  'Sapphire',
+  'Ruby',
+  'Periwinkle',
+  'Beige',
+  'Khaki',
+  'Jade',
+  'Auburn',
+  'Burgundy',
+  'Chartreuse',
+  'Azure',
+  'Ivory',
+  'Champagne',
+  'Taupe',
+  'Orchid',
+  'Plum',
+  'Rose',
+  'Sienna',
+  'Mustard',
+  'Charcoal',
+  'Apricot',
+  'Turmeric',
+  'Saffron',
+  'Lemon',
+  'Raspberry',
+  'Scarlet',
+  'Wine',
+  'Umber',
+  'Cobalt',
+  'Copper',
+  'Sangria',
+  'Caramel',
+  'Sand',
+  'Pistachio',
+  'Lilac',
+];
