+++ apps/ai/src/lib/operations/user/create-personal-team.ts
@@ -3,6 +3,8 @@ import slugify from 'slugify';
 
 import { db } from '#/lib/db';
 
+import { createImage } from './update-new-user';
+
 export const createPersonalTeam = async (userId: string) => {
   const existingPersonalTeam = await db.team.findFirst({
     where: {
@@ -30,14 +32,17 @@ export const createPersonalTeam = async (userId: string) => {
     throw new Error('User not found or displayName is missing');
   }
 
+  const teamImageUrl = createImage(userId, 'team');
+
   const teamSlug = user.name
     ? slugify(user.name, { lower: true, strict: true })
     : user.displayName;
 
   const team = await db.team.create({
     data: {
-      name: user.name ?? user.displayName,
+      name: user.displayName ?? user.name,
       slug: teamSlug,
+      image: teamImageUrl,
       isPersonal: true,
       members: {
         create: {
