--- apps/ai/src/lib/operations/user/generate-unique-display-name.ts
@@ -1,259 +1,0 @@
-import type { Config } from 'unique-names-generator';
-
-import { PrismaClient } from '@prisma/client';
-import { NumberDictionary, uniqueNamesGenerator } from 'unique-names-generator';
-
-const db = new PrismaClient();
-
-const pokemon: string[] = [
-  'Bulbasaur',
-  'Ivysaur',
-  'Venusaur',
-  'Charmander',
-  'Charmeleon',
-  'Charizard',
-  'Squirtle',
-  'Wartortle',
-  'Blastoise',
-  'Caterpie',
-  'Metapod',
-  'Butterfree',
-  'Weedle',
-  'Kakuna',
-  'Beedrill',
-  'Pidgey',
-  'Pidgeotto',
-  'Pidgeot',
-  'Rattata',
-  'Raticate',
-  'Spearow',
-  'Fearow',
-  'Ekans',
-  'Arbok',
-  'Pikachu',
-  'Raichu',
-  'Sandshrew',
-  'Sandslash',
-  'Nidoran',
-  'Nidorina',
-  'Nidoqueen',
-  'Nidoran',
-  'Nidorino',
-  'Nidoking',
-  'Clefairy',
-  'Clefable',
-  'Vulpix',
-  'Ninetales',
-  'Jigglypuff',
-  'Wigglytuff',
-  'Zubat',
-  'Golbat',
-  'Oddish',
-  'Gloom',
-  'Vileplume',
-  'Paras',
-  'Parasect',
-  'Venonat',
-  'Venomoth',
-  'Diglett',
-  'Dugtrio',
-  'Meowth',
-  'Persian',
-  'Psyduck',
-  'Golduck',
-  'Mankey',
-  'Primeape',
-  'Growlithe',
-  'Arcanine',
-  'Poliwag',
-  'Poliwhirl',
-  'Poliwrath',
-  'Abra',
-  'Kadabra',
-  'Alakazam',
-  'Machop',
-  'Machoke',
-  'Machamp',
-  'Bellsprout',
-  'Weepinbell',
-  'Victreebel',
-  'Tentacool',
-  'Tentacruel',
-  'Geodude',
-  'Graveler',
-  'Golem',
-  'Ponyta',
-  'Rapidash',
-  'Slowpoke',
-  'Slowbro',
-  'Magnemite',
-  'Magneton',
-  'Farfetched',
-  'Doduo',
-  'Dodrio',
-  'Seel',
-  'Dewgong',
-  'Grimer',
-  'Muk',
-  'Shellder',
-  'Cloyster',
-  'Gastly',
-  'Haunter',
-  'Gengar',
-  'Onix',
-  'Drowzee',
-  'Hypno',
-  'Krabby',
-  'Kingler',
-  'Voltorb',
-  'Electrode',
-  'Exeggcute',
-  'Exeggutor',
-  'Cubone',
-  'Marowak',
-  'Hitmonlee',
-  'Hitmonchan',
-  'Lickitung',
-  'Koffing',
-  'Weezing',
-  'Rhyhorn',
-  'Rhydon',
-  'Chansey',
-  'Tangela',
-  'Kangaskhan',
-  'Horsea',
-  'Seadra',
-  'Goldeen',
-  'Seaking',
-  'Staryu',
-  'Starmie',
-  'Mr. Mime',
-  'Scyther',
-  'Jynx',
-  'Electabuzz',
-  'Magmar',
-  'Pinsir',
-  'Tauros',
-  'Magikarp',
-  'Gyarados',
-  'Lapras',
-  'Ditto',
-  'Eevee',
-  'Vaporeon',
-  'Jolteon',
-  'Flareon',
-  'Porygon',
-  'Omanyte',
-  'Omastar',
-  'Kabuto',
-  'Kabutops',
-  'Aerodactyl',
-  'Snorlax',
-  'Articuno',
-  'Zapdos',
-  'Moltres',
-  'Dratini',
-  'Dragonair',
-  'Dragonite',
-  'Mewtwo',
-  'Mew',
-];
-
-const colors: string[] = [
-  'Red',
-  'Blue',
-  'Green',
-  'Yellow',
-  'Orange',
-  'Purple',
-  'Magenta',
-  'Turquoise',
-  'Teal',
-  'Lime',
-  'Coral',
-  'Aqua',
-  'Mauve',
-  'Lavender',
-  'Pink',
-  'Maroon',
-  'Olive',
-  'Navy',
-  'Grey',
-  'Tan',
-  'Crimson',
-  'Fuchsia',
-  'Cyan',
-  'Indigo',
-  'Violet',
-  'Cerise',
-  'Rust',
-  'Gold',
-  'Silver',
-  'Bronze',
-  'Amber',
-  'Peach',
-  'Mint',
-  'Emerald',
-  'Sapphire',
-  'Ruby',
-  'Periwinkle',
-  'Beige',
-  'Khaki',
-  'Jade',
-  'Auburn',
-  'Burgundy',
-  'Chartreuse',
-  'Azure',
-  'Ivory',
-  'Champagne',
-  'Taupe',
-  'Orchid',
-  'Plum',
-  'Rose',
-  'Sienna',
-  'Mustard',
-  'Charcoal',
-  'Apricot',
-  'Turmeric',
-  'Saffron',
-  'Lemon',
-  'Raspberry',
-  'Scarlet',
-  'Wine',
-  'Umber',
-  'Cobalt',
-  'Copper',
-  'Sangria',
-  'Caramel',
-  'Sand',
-  'Pistachio',
-  'Lilac',
-];
-
-const config: Config = {
-  dictionaries: [
-    pokemon,
-    colors,
-    NumberDictionary.generate({ min: 0, max: 999 }),
-  ],
-  style: 'lowerCase',
-  separator: '-',
-};
-
-export async function generateUniqueDisplayName(): Promise<string> {
-  let unique = false;
-  let newName = '';
-
-  while (!unique) {
-    newName = uniqueNamesGenerator(config);
-    const existingUser = await db.user.findFirst({
-      where: { displayName: newName },
-    });
-
-    if (!existingUser) {
-      unique = true;
-    }
-  }
-
-  return newName;
-}
